chore: change type of id in table

+++ src/main/java/com/example/demo/domain/GreetingRepository.java
@@ -2,5 +2,5 @@ package com.example.demo.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GreetingRepository extends JpaRepository<Greeting, Long> {
+public interface GreetingRepository extends JpaRepository<Greeting, Integer> {
 }
